Update plugin so it works with Autopsy 4.19.1
Move Language class into main plugin script
add necessary import to plugin
remove Language.py
